tests/cards/base/GeneRepair.spec.ts
import {expect} from 'chai';
import {GeneRepair} from '../../../src/cards/base/GeneRepair';
import {Game} from '../../../src/Game';
import {TestPlayer} from '../../TestPlayer';
import {Resources} from '../../../src/common/Resources';
import {TestPlayers} from '../../TestPlayers';

describe('GeneRepair', function() {
  let card : GeneRepair; let player : TestPlayer;

  beforeEach(function() {
    card = new GeneRepair();
    player = TestPlayers.BLUE.newPlayer();
    const redPlayer = TestPlayers.RED.newPlayer();
    Game.newInstance('foobar', [player, redPlayer], player);
  });

  it('Can\'t play', function() {
    expect(player.canPlayIgnoringCost(card)).is.not.true;
  });

  it('Should play', function() {
    player.playedCards.push(card, card, card);
    expect(player.canPlayIgnoringCost(card)).is.true;
    card.play(player);

    expect(player.getProduction(Resources.MEGACREDITS)).to.eq(2);
    expect(card.getVictoryPoints()).to.eq(2);
  });
});
